% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_operations.R
\name{archive_tasks}
\alias{archive_tasks}
\title{Write A File With To Dos}
\usage{
archive_tasks()
}
\arguments{
\item{foo}{}
}
\value{
bar
}
\description{
empty
}
\author{
Paul Lemmens (paul.lemmens@gmail.com)
}
