% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_time_series.R
\name{complete_years}
\alias{complete_years}
\title{Complete Years Time Series}
\usage{
complete_years(dfr, year_col = year)
}
\arguments{
\item{dfr}{A data frame with a column for year numbers.}

\item{year_col}{A bare column name (defaults to year) to expand to a full
sequence of year numbers.}
}
\value{
The expanded \code{dfr} with a column variable \code{year_col} that contains
a full sequence of year numbers.
}
\description{
Fills gaps in time series data of years that might occur due to missing
data or dates in task lists.
}
