% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{find_tag}
\alias{find_tag}
\title{Find Index of Tag in List of Tags}
\usage{
find_tag(tag_list, tag)
}
\arguments{
\item{tag_list}{A list of strings or vectors of strings representing tasks'
tags (with vectors for tasks with multiple tags).}

\item{tag}{Simple string of one tag to find in the \code{tag_list}.}
}
\value{
A boolean vector for the list indices that contain \code{tag}.
}
\description{
Finds the index of a particular tag in a list of tags wherein the
list might also contain vectors. This function helps to find row
indices for later filtering functions.
}
