% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_by.R
\name{group_by_context}
\alias{group_by_context}
\title{Group Tasks by Their Context}
\usage{
group_by_context(dfr, ...)
}
\description{
Note that grouping by multiple (types of) tags is
accomplished by chaining \code{\link{group_by_context}} and
\code{\link{group_by_project}}.
}
