% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{list_tags}
\alias{list_tags}
\title{List All Tags}
\usage{
list_tags(tasks, type)
}
\arguments{
\item{tasks}{A tibble with tasks as produced by \code{\link{parse_tasks}}. \strong{FIXME}}

\item{type}{A string to indicate whether to list contexts or lists.
Should be one of \code{c('context', '@', 'list', 'project', '+')}.}
}
\value{
A vector of (cleaned) tags.
}
\description{
The todo.txt specification uses \verb{@...} to refer to contexts
and \code{+...} for lists. This function extracts all (unique) contexts
or lists from the todo.txt file.
}
