% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_operations.R
\name{load_tasks}
\alias{load_tasks}
\title{Open A File With To Dos}
\usage{
load_tasks(f)
}
\arguments{
\item{f}{An absolute path to a todo.txt file (string).}
}
\value{
A tibble with the to dos loaded from file with variables as
specified in \code{\link[=parse_prefixes]{parse_prefixes()}}.
}
\description{
This function loads tasks from a file provided. This file typically is
the \code{todo.txt} file but can also be a \code{done.txt} file.
}
\author{
Paul Lemmens (paul.lemmens@gmail.com)
}
