% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{modify_tag}
\alias{modify_tag}
\title{Modify Task Tag(s)}
\usage{
modify_tag(task, old, new)
}
\arguments{
\item{task}{A string with one (or, as vector, one or more) task(s) that
follow(s) the todo.txt specification by Gina Trapani. When multiple
tasks are provided, \emph{all} tasks are modified in the same way.}

\item{old}{Old tag to find and replace (a string).}

\item{new}{New tag to replace the old value with. Note, could be multiple
new tags chained together \code{+new1 +new2} (a single string).}
}
\value{
the updated task(s)
}
\description{
This is a package internal function to modify tags of a task (in the
todo.txt definition properties prefixed with \code{+} or \code{@}).

Note that by suitably setting the \code{old} and \code{new} value, specific effects
like removing a tag or adding an additional tag. For instance, by setting
the \code{old} value to the end-of-line regular expression \code{$}, a tag can be
appended. Note that this may result in adding superfluous spaces.
}
\author{
Paul Lemmens (paul.lemmens@gmail.com)
}
