% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{parse_prefixes}
\alias{parse_prefixes}
\title{Parse Priority, Creation And Completed Dates}
\usage{
parse_prefixes(todo)
}
\arguments{
\item{todo}{A to do (string), or a vector of to dos.}
}
\value{
A tibble with variables for creation date, completion date, and
priority of a to do
}
\description{
Parsing a to do's priority, creation and due date actually combines
elements that should be handled separately. However, the grep pattern
that is used is quite useful and more difficult to strip into components.
Additionally, these components are all prefixed before the actual task
so it is sensible to combine parsing in one go.
}
\references{
https://github.com/mpcjanssen/simpletask-android/blob/197bd51f496bd6066df902445acc28df51910d60/src/main/java/nl/mpcjanssen/simpletask/task/Task.java
}
\author{
Paul Lemmens (paul.lemmens@gmail.com)
}
