% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{parse_tags}
\alias{parse_tags}
\title{Parse Tags From A Task}
\usage{
parse_tags(todo, tag)
}
\arguments{
\item{todo}{A single todo (string).}

\item{tag}{A plus \code{+} or \code{@} sign (string)}
}
\value{
A (mini) data frame with a row for each context found and in that
row a variable for the context and its start and end position in the
task.
}
\description{
Contexts are tags prefixed with an \code{@}. Projects are tags prefixed with
a \code{+} (plus). One task can contain multiple
contexts and/or multiple projects that should all be preserved.
The regular expressions are taken from the source code of Simpletask by
Mark Janssen.
}
\author{
Paul Lemmens (paul.lemmens@gmail.com)
}
