% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{parse_recurrence}
\alias{parse_recurrence}
\title{Parse Recurrence In A Task}
\usage{
parse_recurrence(todo)
}
\arguments{
\item{todo}{A single task (string) or vector of tasks.}
}
\value{
The recurrence setting for a task as string.
}
\description{
A further extension of the original specification is setting a
recurrence of a task using the prefix \code{rec:}. A recurrence is set in
N days, weeks, months, or years. When N is preceded by a plus, it says to
start the recurrence calculation by the original due date or by today's
date.
}
\author{
Paul Lemmens (paul.lemmens@gmail.com)
}
