% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{parse_task}
\alias{parse_task}
\alias{parse_tasks}
\title{Parse A (List Of) Tasks}
\usage{
parse_task(task)

parse_tasks(task)
}
\arguments{
\item{task}{One task (string), or a vector of strings.}
}
\value{
A tibble with a parsed task.
}
\description{
This function parses one task, or a "list" (read, a vector of task
strings) in its entirety. A task is specified following the definition(s)
by Gina Trapani via todotxt.org and comprising a creation date as first
item on a line, a due date marked using \verb{due:}, and a priority at the start
of each line as \code{(X)}. Tags are indicated using either an \code{@} for
(typically) contexts and a plus \code{+} for projects. Optional components
comprise a recurrence specified
using \verb{rec:} as extension of the original todo.txt specification (see
documentation of the android app SimpleTask).
}
\author{
Paul Lemmens (paul.lemmens@paul-lemmens.nl)
}
